requirements.txt
defusedxml~=0.7.1
ipython~=8.23.0
numpy~=1.26.4
pillow~=10.3.0
pip~=25.1.1
attrs~=23.2.0
wheel~=0.43.0
pytz~=2024.1
PyYAML~=6.0.1
black~=25.1.0
platformdirs~=4.2.0
colorama~=0.4.6
mypy_extensions~=1.1.0
packaging~=24.0
pathspec~=0.12.1
click~=8.2.1
setuptools~=69.5.1
Django~=5.2.4
asgiref~=3.9.0
sqlparse~=0.5.3
Jinja2~=3.1.3
requests~=2.31.0
Pygments~=2.17.2
wcwidth~=0.2.13
djangorestframework~=3.16.0
inflection~=0.5.1
uritemplate~=4.2.0
referencing~=0.34.0
jsonschema~=4.21.1
idna~=3.7
webcolors~=1.13
jsonpointer~=2.4
uri-template~=1.3.0
isoduration~=20.11.0
fqdn~=1.5.1
rfc3339-validator~=0.1.4
jsonschema-specifications~=2023.12.1
PyJWT~=2.10.1
drf-yasg~=1.21.10
django-filter~=25.1
django-perms~=2.1.0
mypy_extensions~=1.1.0
typing_extensions~=4.14.1
psycopg~=3.2.9
psycopg2-binary~=2.9.10
django-cors-headers~=4.7.0
jedi~=0.19.1